Правки.

+++ src/main/java/by/forward/forward_system/core/dto/ui/ReviewDto.java
@@ -15,7 +15,7 @@ public class ReviewDto {
     private Long attachmentId;
     private String reviewMessage;
     private String reviewVerdict;
-    private Integer reviewVerdictMark;
+    private Double reviewVerdictMark;
     private Long reviewFileId;
     private Boolean isReviewed;
     private Boolean isAccepted;
